make adjustments to files/structure for consistency - add .prettierignore files - update .tsconfigs - move JetIcon inside an "icon" directory - tweak App.tsx in both templates - change ApolloWrapper and MainNavigator to named exports in jet - get rid of unused "NavigationWrapper" in jet
