+++ templates/blimp/constants.ts
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/camelcase */
+import Config from 'react-native-config';
+
+export const API_BASE_URL = Config.API_BASE_URL || '';
