+++ templates/jet/app/apollo/ApolloWrapper.tsx
@@ -3,10 +3,8 @@ import { ApolloProvider } from '@apollo/client';
 
 import { useApolloClient } from './ApolloUtils';
 
-const ApolloWrapper: React.FC = ({ children }) => {
+export const ApolloWrapper: React.FC = ({ children }) => {
   const client = useApolloClient();
 
   return <ApolloProvider client={client}>{children}</ApolloProvider>;
 };
-
-export default ApolloWrapper;
