+++ templates/jet/app/assets/icons/JetIcon.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { StyleProp, ViewStyle } from 'react-native';
 import { Svg, G, Path } from 'react-native-svg';
 
-import { View } from '../components/common/View';
-import { colors } from '../styles/globalStyles';
+import { View } from '../../components/common/View';
+import { colors } from '../../styles/globalStyles';
 
 type Props = {
   containerStyle?: StyleProp<ViewStyle>;
